+++ apps/sif-ettersending/src/app/pages/velkommen-page/VelkommenPage.tsx
@@ -1,38 +1,33 @@
-import React from 'react';
 import { SIFCommonPageKey, useLogSidevisning } from '@navikt/sif-common-amplitude';
-import Page from '@navikt/sif-common-core-ds/src/components/page/Page';
-import { SamtykkeForm, SoknadHeader } from '@navikt/sif-common-soknad-ds';
-import { useAppIntl } from '../../i18n';
+import { SoknadVelkommenPage } from '@navikt/sif-common-soknad-ds';
+import { AppText, useAppIntl } from '../../i18n';
 import { useSoknadContext } from '../../soknad/SoknadContext';
+import { Person } from '../../types/Person';
 import { Søknadstype } from '../../types/Søknadstype';
 import OmSøknaden from './OmSøknaden';
-import VelkommenGuide from './VelkommenGuide';
 
 interface Props {
     søknadstype: Søknadstype;
+    søker: Person;
 }
 
-const VelkommenPage: React.FC<Props> = ({ søknadstype }) => {
+const VelkommenPage = ({ søknadstype, søker }: Props) => {
     const { text } = useAppIntl();
     const { startSoknad } = useSoknadContext();
 
     useLogSidevisning(SIFCommonPageKey.velkommen);
 
     return (
-        <>
-            <Page
-                title={text(`application.title.${søknadstype}`)}
-                topContentRenderer={() => <SoknadHeader title={text(`application.title.${søknadstype}`)} level="2" />}>
-                <VelkommenGuide />
-
-                <OmSøknaden />
-
-                <SamtykkeForm
-                    onValidSubmit={startSoknad}
-                    submitButtonLabel={text('ettersendelse.samtykkeForm.submitButtonLabel')}
-                />
-            </Page>
-        </>
+        <SoknadVelkommenPage
+            title={text(`application.title.${søknadstype}`)}
+            onStartSøknad={startSoknad}
+            submitButtonLabel={text('ettersendelse.samtykkeForm.submitButtonLabel')}
+            guide={{
+                navn: søker.fornavn,
+                content: <AppText id="page.velkommen.guide.ingress" />,
+            }}>
+            <OmSøknaden />
+        </SoknadVelkommenPage>
     );
 };